requirements.txt
fitz==0.0.1.dev2
langchain_core==1.0.0
langchain_openai==1.0.0
pdf2image==1.17.0
pdfplumber==0.11.7
Pillow==12.0.0
pymilvus==2.6.2
Requests==2.32.5
